Move unit test for preprocessing output into row preprocessing alg
Re #32125
