+++ qt/scientific_interfaces/ISISReflectometry/test/Preview/PreviewPresenterTest.h
@@ -30,18 +30,6 @@ using ::testing::Return;
 using ::testing::ReturnRef;
 
 class PreviewPresenterTest : public CxxTest::TestSuite {
-  class StubbedPreProcess : public WorkspaceCreationHelper::StubAlgorithm {
-  public:
-    StubbedPreProcess() { this->setChild(true); }
-    void addOutputWorkspace(Mantid::API::MatrixWorkspace_sptr &ws) {
-      const std::string propName = "OutputWorkspace";
-      auto prop = std::make_unique<Mantid::API::WorkspaceProperty<>>(propName, "", Mantid::Kernel::Direction::Output);
-      prop->createTemporaryValue();
-      declareProperty(std::move(prop));
-      setProperty(propName, ws);
-    }
-  };
-
 public:
   void test_notify_load_workspace_requested() {
     auto mockModel = std::make_unique<MockPreviewModel>();
@@ -57,23 +45,6 @@ public:
   }
 
   void test_notify_preprocessing_algorithm_complete() {
-    auto mockModel = std::make_unique<MockPreviewModel>();
-    auto mockView = std::make_unique<MockPreviewView>();
-
-    auto mockAlg = std::make_shared<StubbedPreProcess>();
-    const bool isHistogram = true;
-    Mantid::API::MatrixWorkspace_sptr mockWs = WorkspaceCreationHelper::create1DWorkspaceRand(1, isHistogram);
-    mockAlg->addOutputWorkspace(mockWs);
-
-    auto properties = IConfiguredAlgorithm::AlgorithmRuntimeProps();
-    auto runNumbers = std::vector<std::string>{};
-    auto row = PreviewRow(runNumbers);
-    auto configuredAlg = std::make_shared<BatchJobAlgorithm>(std::static_pointer_cast<Mantid::API::IAlgorithm>(mockAlg),
-                                                             properties, nullptr, &row);
-
-    auto presenter = PreviewPresenter(mockView.get(), std::move(mockModel));
-    presenter.notifyAlgorithmComplete(configuredAlg);
-
-    // TODO ASSERT WS_Ptr
+    // TODO
   }
 };
